requirements.txt
click>=8.0.0
pillow>=9.0.0
jinja2>=3.0.0
pytest>=7.0.0
